+++ crates/amaru-protocols/src/tests/slow_manager_stage.rs
@@ -39,6 +39,7 @@ pub async fn slow_manager_stage(manager: Manager, msg: ManagerMessage, eff: Effe
         ManagerMessage::NewTip(_, _) => {}
         ManagerMessage::FetchBlocks { .. } => {}
         ManagerMessage::RequestSharePeers { .. } => {}
+        ManagerMessage::ShareRequest { .. } => {}
         ManagerMessage::ConnectionResult(..) => {}
     }
     manager::stage(manager, msg, eff).await